App structure and components optimisation
