+++ next.config.js
@@ -4,7 +4,7 @@ const nextConfig = {
     return [
       {
         source: '/api/:path*',
-        destination: 'http://localhost:3000/api/:path*',
+        destination: process.env.NEXT_PUBLIC_API_URL + 'api/:path*',
         basePath: false,
       },
     ]
