+++ src/pages/index.tsx
@@ -1,13 +1,6 @@
 import { useState, useEffect, FC, PropsWithChildren } from 'react'
 import { useRouter } from 'next/router'
-import {
-  ChevronDown,
-  ChevronUp,
-  AlertTriangle,
-  Clock,
-  CheckCircle2,
-  MessageSquare,
-} from 'lucide-react'
+import { ChevronDown, ChevronUp, AlertTriangle } from 'lucide-react'
 import {
   Select,
   SelectContent,
@@ -17,6 +10,8 @@ import {
 } from '@/components/select/Select'
 import { Usuario } from '@/types/Usuario'
 import { Interconsulta } from '@/types/Interconsulta'
+import { Servicio } from '@/types/Servicio'
+import InterconsultaCard from '@/components/interconsulta-card/InterconsultaCard'
 
 type CollapsibleSectionProps = {
   title: string
@@ -66,31 +61,17 @@ const VerInterconsultas = () => {
   const [loading, setLoading] = useState(true)
   const [error, setError] = useState<string | undefined>(undefined)
   const [usuario, setUsuario] = useState<Usuario | undefined>(undefined)
-  const [servicios, setServicios] = useState<any[]>([])
+  const [servicios, setServicios] = useState<Servicio[]>([])
   const [filtros, setFiltros] = useState({
     estado: '',
     prioridad: '',
     servicio: '',
   })
 
-  const formatSignoVitalLabel = (key: string) => {
-    const labels: { [key: string]: string } = {
-      presionArterial: 'Presión Arterial',
-      frecuenciaCardiaca: 'Frecuencia Cardíaca',
-      frecuenciaRespiratoria: 'Frecuencia Respiratoria',
-      temperatura: 'Temperatura',
-      saturacionOxigeno: 'Saturación de Oxígeno',
-    }
-    return labels[key] || key
-  }
-
   const fetchServicios = async () => {
     try {
       const token = window.localStorage.getItem('token')
       const apiUrl = `${process.env.NEXT_PUBLIC_API_URL}/api/servicios`
-
-      console.log('Fetching services from:', apiUrl)
-      console.log('Using token:', token)
 
       const response = await fetch(apiUrl, {
         headers: {
@@ -160,6 +141,8 @@ const VerInterconsultas = () => {
       const dataEnviadas = await responseEnviadas.json()
       const dataRecibidas = await responseRecibidas.json()
 
+      console.log(dataEnviadas, dataRecibidas)
+
       setInterconsultasEnviadas(
         Array.isArray(dataEnviadas.data) ? dataEnviadas.data : []
       )
@@ -198,326 +181,6 @@ const VerInterconsultas = () => {
       fetchInterconsultas()
     }
   }, [filtros, usuario])
-
-  type InterconsultaCardProps = {
-    interconsulta: Interconsulta
-    onStatusChange: () => void
-  }
-  const InterconsultaCard: FC<InterconsultaCardProps> = ({
-    interconsulta,
-    onStatusChange,
-  }) => {
-    const [expanded, setExpanded] = useState(false)
-    const [updating, setUpdating] = useState(false)
-    const [updateError, setUpdateError] = useState<string>('')
-
-    const handleRespuestaVirtual = (e: any) => {
-      e.preventDefault()
-      e.stopPropagation()
-      router.push(`/interconsultas/${interconsulta._id}/respuesta-virtual`)
-    }
-
-    const handleRespuestaFisica = (e: any) => {
-      e.preventDefault()
-      e.stopPropagation()
-      console.log('Respuesta Física clickeada')
-    }
-
-    if (
-      loading &&
-      interconsultasEnviadas.length === 0 &&
-      interconsultasRecibidas.length === 0
-    ) {
-      return (
-        <div className="min-h-screen bg-gray-50">
-          <div className="container mx-auto p-4">
-            <div className="text-center py-6">Cargando interconsultas...</div>
-          </div>
-        </div>
-      )
-    }
-
-    if (error) {
-      return (
-        <div className="min-h-screen bg-gray-50">
-          <div className="container mx-auto p-4">
-            <div className="bg-red-50 border border-red-200 text-red-800 px-4 py-3 rounded-lg">
-              <div className="flex items-center">
-                <AlertTriangle className="h-5 w-5 mr-2" />
-                <span>{error}</span>
-              </div>
-            </div>
-          </div>
-        </div>
-      )
-    }
-
-    const handleStatusChange = async (newStatus: string) => {
-      if (updating) return
-
-      try {
-        setUpdating(true)
-        setUpdateError('')
-        const token = window.localStorage.getItem('token')
-        const response = await fetch(
-          `${process.env.NEXT_PUBLIC_API_URL}/api/interconsultas/${interconsulta._id}/estado`,
-          {
-            method: 'PUT',
-            headers: {
-              'Content-Type': 'application/json',
-              Authorization: `Bearer ${token}`,
-            },
-            body: JSON.stringify({ estado: newStatus }),
-          }
-        )
-
-        if (!response.ok) {
-          throw new Error('Error al actualizar el estado')
-        }
-
-        const data = await response.json()
-        if (data.exito) {
-          onStatusChange()
-        }
-      } catch (error) {
-        console.error('Error:', error)
-        setUpdateError('Error al actualizar el estado')
-      } finally {
-        setUpdating(false)
-      }
-    }
-
-    const getStatusColor = (estado: string) => {
-      const colors: { [key: string]: string } = {
-        PENDIENTE: 'bg-yellow-100 text-yellow-800 border-yellow-200',
-        EN_PROCESO: 'bg-blue-100 text-blue-800 border-blue-200',
-        COMPLETADA: 'bg-green-100 text-green-800 border-green-200',
-      }
-      return colors[estado] || 'bg-gray-100 text-gray-800 border-gray-200'
-    }
-
-    const getPriorityIcon = (prioridad: string) => {
-      switch (prioridad) {
-        case 'ALTA':
-          return <AlertTriangle className="h-5 w-5 text-red-500" />
-        case 'MEDIA':
-          return <Clock className="h-5 w-5 text-yellow-500" />
-        case 'BAJA':
-          return <CheckCircle2 className="h-5 w-5 text-green-500" />
-        default:
-          return null
-      }
-    }
-
-    const formatFecha = (fecha: string) => {
-      try {
-        return new Date(fecha).toLocaleDateString('es-ES', {
-          weekday: 'long',
-          year: 'numeric',
-          month: 'long',
-          day: 'numeric',
-          hour: '2-digit',
-          minute: '2-digit',
-        })
-      } catch (error) {
-        return 'Fecha inválida'
-      }
-    }
-
-    if (!interconsulta || !interconsulta.paciente) {
-      return null
-    }
-
-    return (
-      <div className="bg-white rounded-lg shadow-sm hover:shadow-md transition-all duration-300 border border-gray-100">
-        <div className="p-4">
-          <div
-            className="cursor-pointer select-none"
-            onClick={() => setExpanded(!expanded)}
-          >
-            <div className="flex justify-between items-start">
-              <div className="space-y-2">
-                <div className="flex items-center space-x-2">
-                  <h2 className="text-black font-semibold">
-                    {interconsulta.paciente?.nombre}
-                  </h2>
-                  {getPriorityIcon(interconsulta.prioridad)}
-                </div>
-                <div className="space-y-1">
-                  <p className="text-sm text-gray-800">
-                    HC: {interconsulta.paciente?.numeroHistoria}
-                  </p>
-                  <p className="text-sm text-gray-800">
-                    De: {interconsulta.servicioSolicitante?.nombre}
-                  </p>
-                  <p className="text-sm text-gray-800">
-                    Para: {interconsulta.servicioDestino?.nombre}
-                  </p>
-                </div>
-              </div>
-              <div className="flex flex-col items-end gap-2">
-                <span
-                  className={`px-3 py-1 rounded-full text-sm border ${getStatusColor(
-                    interconsulta.estado
-                  )}`}
-                >
-                  {interconsulta.estado}
-                </span>
-                <span className="text-xs text-gray-500">
-                  {formatFecha(interconsulta.fechaCreacion)}
-                </span>
-                {expanded ? (
-                  <ChevronUp className="h-5 w-5 text-gray-400" />
-                ) : (
-                  <ChevronDown className="h-5 w-5 text-gray-400" />
-                )}
-              </div>
-            </div>
-          </div>
-
-          {expanded && (
-            <div className="mt-4 pt-4 border-t border-gray-100">
-              <div className="grid gap-6">
-                {/* Buttons section - only show for EN_PROCESO status */}
-                {interconsulta.estado === 'EN_PROCESO' && (
-                  <div className="border-b border-gray-100 pb-4">
-                    <div className="flex justify-end space-x-4">
-                      <button
-                        type="button"
-                        className="flex items-center gap-2 px-4 py-2 bg-gray-100 text-gray-700 rounded-md hover:bg-gray-200 transition-colors"
-                        onClick={handleRespuestaFisica}
-                      >
-                        <CheckCircle2 className="h-4 w-4" />
-                        Respuesta Física
-                      </button>
-                      <button
-                        type="button"
-                        className="flex items-center gap-2 px-4 py-2 bg-blue-500 text-white rounded-md hover:bg-blue-600 transition-colors"
-                        onClick={handleRespuestaVirtual}
-                      >
-                        <MessageSquare className="h-4 w-4" />
-                        Respuesta Virtual
-                      </button>
-                    </div>
-                  </div>
-                )}
-
-                <div className="flex items-center justify-between bg-gray-50 p-4 rounded-lg border border-gray-200">
-                  <div className="space-y-1">
-                    <h3 className="text-sm font-semibold text-gray-900">
-                      Estado de la Interconsulta
-                    </h3>
-                    <p className="text-sm text-gray-800">
-                      Estado actual: {interconsulta.estado}
-                    </p>
-                    {updateError && (
-                      <p className="text-sm text-red-600">{updateError}</p>
-                    )}
-                  </div>
-                  <select
-                    value={interconsulta.estado}
-                    onChange={(e) => handleStatusChange(e.target.value)}
-                    disabled={updating}
-                    className={`
-                      rounded-md border border-gray-300 px-3 py-2 text-gray-700 
-                      focus:outline-none focus:ring-2 focus:ring-blue-500
-                      ${updating ? 'opacity-50 cursor-not-allowed' : ''}
-                    `}
-                  >
-                    <option value="PENDIENTE">Pendiente</option>
-                    <option value="EN_PROCESO">En Proceso</option>
-                    <option value="COMPLETADA">Completada</option>
-                  </select>
-                </div>
-
-                <div>
-                  <h3 className="text-sm font-semibold text-gray-900 mb-2">
-                    Objetivo de la Consulta
-                  </h3>
-                  <p className="text-sm text-gray-900 bg-gray-50 p-3 rounded">
-                    {interconsulta.objetivoConsulta}
-                  </p>
-                </div>
-
-                {interconsulta.estadoClinico && (
-                  <div className="space-y-4">
-                    <div>
-                      <h3 className="text-sm font-semibold text-gray-900 mb-2">
-                        Estado Clínico
-                      </h3>
-                      <p className="text-sm text-gray-900 bg-gray-50 p-3 rounded">
-                        {interconsulta.estadoClinico.subjetivo}
-                      </p>
-                    </div>
-
-                    {interconsulta.estadoClinico.signosVitales && (
-                      <div>
-                        <h3 className="text-sm font-semibold text-gray-900 mb-2">
-                          Signos Vitales
-                        </h3>
-                        <div className="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-3">
-                          {Object.entries(
-                            interconsulta.estadoClinico.signosVitales
-                          ).map(([key, value]) => (
-                            <div key={key} className="bg-gray-50 p-3 rounded">
-                              <p className="text-xs text-gray-700 mb-1">
-                                {formatSignoVitalLabel(key)}
-                              </p>
-                              <p className="text-sm font-medium text-black">
-                                {value}
-                              </p>
-                            </div>
-                          ))}
-                        </div>
-                      </div>
-                    )}
-                  </div>
-                )}
-
-                {(interconsulta.laboratorios || interconsulta.imagenologia) && (
-                  <div className="grid md:grid-cols-2 gap-4">
-                    {interconsulta.laboratorios && (
-                      <div>
-                        <h3 className="text-sm font-semibold text-gray-900 mb-2">
-                          Laboratorios
-                        </h3>
-                        <div className="bg-gray-50 p-3 rounded space-y-2">
-                          <p className="text-sm text-black">
-                            {interconsulta.laboratorios.resultados}
-                          </p>
-                          {interconsulta.laboratorios.observaciones && (
-                            <p className="text-sm text-black">
-                              Nota: {interconsulta.laboratorios.observaciones}
-                            </p>
-                          )}
-                        </div>
-                      </div>
-                    )}
-
-                    {interconsulta.imagenologia && (
-                      <div>
-                        <h3 className="text-sm font-semibold text-gray-900 mb-2">
-                          Imagenología
-                        </h3>
-                        <div className="bg-gray-50 p-3 rounded space-y-2">
-                          <p className="text-sm font-medium text-black">
-                            {interconsulta.imagenologia.tipo}
-                          </p>
-                          <p className="text-sm text-black">
-                            {interconsulta.imagenologia.hallazgosRelevantes}
-                          </p>
-                        </div>
-                      </div>
-                    )}
-                  </div>
-                )}
-              </div>
-            </div>
-          )}
-        </div>
-      </div>
-    )
-  }
 
   if (
     loading &&
@@ -673,7 +336,7 @@ const VerInterconsultas = () => {
                     value={servicio._id}
                     className="text-gray-900 hover:bg-gray-100 bg-white"
                   >
-                    {servicio.nombre}
+                    {/* {servicio.nombre} */}
                   </SelectItem>
                 ))}
               </SelectContent>
@@ -693,11 +356,15 @@ const VerInterconsultas = () => {
             ) : (
               <div className="space-y-4">
                 {interconsultasEnviadas.map((interconsulta) =>
-                  interconsulta && interconsulta._id ? (
+                  interconsulta ? (
                     <InterconsultaCard
                       key={interconsulta._id}
                       interconsulta={interconsulta}
                       onStatusChange={fetchInterconsultas}
+                      loading={loading}
+                      error={error || ''}
+                      interconsultasEnviadas={interconsultasEnviadas}
+                      interconsultasRecibidas={interconsultasRecibidas}
                     />
                   ) : null
                 )}
@@ -715,15 +382,17 @@ const VerInterconsultas = () => {
               </div>
             ) : (
               <div className="space-y-4">
-                {interconsultasRecibidas.map((interconsulta) =>
-                  interconsulta && interconsulta._id ? (
-                    <InterconsultaCard
-                      key={interconsulta._id}
-                      interconsulta={interconsulta}
-                      onStatusChange={fetchInterconsultas}
-                    />
-                  ) : null
-                )}
+                {interconsultasRecibidas.map((interconsulta) => (
+                  <InterconsultaCard
+                    key={interconsulta._id}
+                    interconsulta={interconsulta}
+                    onStatusChange={fetchInterconsultas}
+                    loading={loading}
+                    error={error || ''}
+                    interconsultasEnviadas={interconsultasEnviadas}
+                    interconsultasRecibidas={interconsultasRecibidas}
+                  />
+                ))}
               </div>
             )}
           </CollapsibleSection>
